# V2.3/code/_openapi_snapshot.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Starriver SuperBrain API",
    "description": "Minimal runnable V2.3 API skeleton",
    "version": "v2.3-preview"
  },
  "paths": {
...
    "/api/v2.3-preview/memory/sync": {
      "post": {
        "tags": [
          "memory"
        ],
        "summary": "Memory Sync",
        "operationId": "memory_sync_api_v2_3_preview_memory_sync_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemorySyncRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemorySyncResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2.3-preview/memory/export": {
      "get": {
        "tags": [
          "memory"
        ],
        "summary": "Memory Export",
        "operationId": "memory_export_api_v2_3_preview_memory_export_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryExportResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "memory"
        ],
        "summary": "Memory Export Post",
        "operationId": "memory_export_post_api_v2_3_preview_memory_export_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemoryExportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryExportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MemoryExportRequest": {
        "properties": {
          "format": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Format",
            "default": "json"
          },
          "limit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Limit",
            "default": 100
          }
        },
        "type": "object",
        "title": "MemoryExportRequest"
      },
      "MemoryExportResponse": {
        "properties": {
          "export_url": {
            "type": "string",
            "title": "Export Url"
          },
          "expires_at": {
            "type": "string",
            "title": "Expires At"
          }
        },
        "type": "object",
        "required": [
          "export_url",
          "expires_at"
        ],
        "title": "MemoryExportResponse"
      },
      "MemoryItem": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "content": {
            "type": "object",
            "title": "Content"
          }
        },
        "type": "object",
        "required": [
          "content"
        ],
        "title": "MemoryItem"
      },
      "MemorySyncRequest": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/MemoryItem"
            },
            "type": "array",
            "title": "Items"
          },
          "force": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Force",
            "default": false
          },
          "timeout": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timeout",
            "default": 30
          }
        },
        "type": "object",
        "title": "MemorySyncRequest"
      },
      "MemorySyncResponse": {
        "properties": {
          "synced_count": {
            "type": "integer",
            "title": "Synced Count"
          },
          "failed_count": {
            "type": "integer",
            "title": "Failed Count"
          },
          "trace_id": {
            "type": "string",
            "title": "Trace Id"
          },
          "finished_at": {
            "type": "string",
            "title": "Finished At"
          }
        },
        "type": "object",
        "required": [
          "synced_count",
          "failed_count",
          "trace_id",
          "finished_at"
        ],
        "title": "MemorySyncResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}